Add files via upload
Edit to the original program to have a method that decides if the user wants to use the normal sort, or the reverse sort. Also added a line to print "Done" when the results are posted so the user can know if the program actually ran properly from the cmd interface.